ai/clip/requirements.txt
# CLIP Embedding Server Dependencies

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Deep learning frameworks
torch>=2.0.0
transformers>=4.35.0

# Image processing
pillow>=10.0.0
numpy>=1.24.0

# Utilities
pydantic>=2.4.0

# Monitoring
pynvml>=11.5.0
